Fix for missing Flow information when paging limit is configured by the user to a value larger than the Registry's default.
The trouble is that a react-admin ReferenceField in a list view results in a GET_MANY request to the data provider, on which we were not adding a `paging.limit` parameter, and the Query API may therefore make a page response with a lower number.

The correct fix is to handle a paged response (similarly to how we do when we don't have RQL support to do GET_MANY), but the quick fix (provided the maximum permitted by the Registry is large enough) is to follow our implementation in GET_MANY_REFERENCE and hard-code a large number.